display individual spritegroups and sprites
