+++ ImasArchiveApp/UIElementImageView.xaml
@@ -1,0 +1,15 @@
+﻿<UserControl
+    x:Class="ImasArchiveApp.UIElementImageView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:ImasArchiveApp"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    d:DesignHeight="720"
+    d:DesignWidth="1280"
+    mc:Ignorable="d">
+    <Grid Width="1280" Height="720">
+        <Image Source="/Icons/checkerboard.png" Stretch="UniformToFill" />
+        <Image Source="{Binding Path=ImageSource}" />
+    </Grid>
+</UserControl>
