+++ ImasArchiveApp/ViewModels/UI/UISpriteModel.cs
@@ -8,6 +8,7 @@ namespace ImasArchiveApp
     class UISpriteModel : UIElementModel
     {
         private readonly Sprite _sprite;
+        protected override UIElement UIElement => _sprite;
         public override string ModelName => $"({_sprite.width}x{_sprite.height})";
         public UISpriteModel(UISubcomponentModel parent, Sprite sprite) : base(parent, "sprite")
         {
